Refactor the requests module
Move the utility functions and staticmethods out of requests module into
a new request_utils module.
