+++ test/unit/test_request.py
@@ -4,14 +4,14 @@ import logging
 
 from functools import partial
 
-from jsonrpcserver.request import Request, _convert_camel_case, \
-    _convert_camel_case_keys
-from jsonrpcserver.response import ErrorResponse, RequestResponse, \
-    NotificationResponse
+from jsonrpcserver.request import Request
+from jsonrpcserver.response import (
+    ErrorResponse, RequestResponse, NotificationResponse)
 from jsonrpcserver.exceptions import InvalidParams, MethodNotFound
 from jsonrpcserver.methods import Methods
 from jsonrpcserver import status
 from jsonrpcserver import config
+
 
 def setUpModule():
     config.debug = True
@@ -20,87 +20,66 @@ def tearDownModule():
     config.debug = False
 
 
+# Some dummy functions to use for testing
 def foo():
     return 'bar'
 
-
-class TestConvertCamelCase(TestCase):
-
-    def test(self):
-        self.assertEqual('foo_bar', _convert_camel_case('fooBar'))
+class FooClass():
+    def foo(self):
+        return 'bar'
 
 
-class TestConvertCamelCaseKeys(TestCase):
+class TestRequestInit(TestCase):
 
-    def test(self):
-        dictionary = {'fooKey': 1, 'aDict': {'fooKey': 1, 'barKey': 2}}
-        self.assertEqual({'foo_key': 1, 'a_dict': {'foo_key': 1, 'bar_key': 2}}, \
-                _convert_camel_case_keys(dictionary))
+    def tearDown(self):
+        config.convert_camel_case = False
 
+    def test_invalid_request(self):
+        req = Request({'jsonrpc': '2.0'})
+        self.assertIsInstance(req.response, ErrorResponse)
 
-class TestValidateArgumentsAgainstSignature(TestCase):
-    """Keep it simple here. No need to test the signature.bind function."""
-
-    @staticmethod
-    def test_no_arguments():
+    def test_ok(self):
         req = Request({'jsonrpc': '2.0', 'method': 'foo'})
-        req._validate_arguments_against_signature(lambda: None)
+        self.assertEqual('foo', req.method_name)
 
-    def test_no_arguments_too_many_positionals(self):
-        req = Request({'jsonrpc': '2.0', 'method': 'foo', 'params': ['foo']})
-        with self.assertRaises(InvalidParams):
-            req._validate_arguments_against_signature(lambda: None)
+    def test_positional_args(self):
+        req = Request({'jsonrpc': '2.0', 'method': 'foo', 'params': [2, 3]})
+        self.assertEqual([2, 3], req.args)
 
-    @staticmethod
-    def test_positionals():
-        req = Request({'jsonrpc': '2.0', 'method': 'foo', 'params': [1]})
-        req._validate_arguments_against_signature(lambda x: None)
-
-    def test_positionals_not_passed(self):
+    def test_keyword_args(self):
         req = Request({'jsonrpc': '2.0', 'method': 'foo', 'params': {'foo': 'bar'}})
-        with self.assertRaises(InvalidParams):
-            req._validate_arguments_against_signature(lambda x: None)
+        self.assertEqual({'foo': 'bar'}, req.kwargs)
 
-    @staticmethod
-    def test_keywords():
-        req = Request({'jsonrpc': '2.0', 'method': 'foo', 'params': {'foo': 'bar'}})
-        req._validate_arguments_against_signature(lambda **kwargs: None)
+    def test_request_id(self):
+        req = Request({'jsonrpc': '2.0', 'method': 'foo', 'id': 99})
+        self.assertEqual(99, req.request_id)
+
+    def test_request_id_notification(self):
+        req = Request({'jsonrpc': '2.0', 'method': 'foo'})
+        self.assertEqual(None, req.request_id)
+
+    def test_convert_camel_case(self):
+        config.convert_camel_case = True
+        req = Request({'jsonrpc': '2.0', 'method': 'fooMethod', 'params': {
+            'fooParam': 1, 'aDict': {'barParam': 1}}})
+        self.assertEqual('foo_method', req.method_name)
+        self.assertEqual({'foo_param': 1, 'a_dict': {'bar_param': 1}}, req.kwargs)
+
+    def test_positional_args_convert_case_skip(self):
+        config.convert_camel_case = True
+        req = Request({'jsonrpc': '2.0', 'method': 'foo', 'params': ['Camel', 'Case']})
+        self.assertEqual(['Camel', 'Case'], req.args)
 
 
+class TestRequestIsNotification(TestCase):
 
-class TestGetMethod(TestCase):
+    def test_true(self):
+        req = Request({'jsonrpc': '2.0', 'method': 'foo'})
+        self.assertTrue(req.is_notification)
 
-    def test_list(self):
-        def cat(): pass
-        def dog(): pass
-        self.assertIs(cat, Request._get_method([cat, dog], 'cat'))
-        self.assertIs(dog, Request._get_method([cat, dog], 'dog'))
-
-    def test_list_non_existant(self):
-        def cat(): pass
-        with self.assertRaises(MethodNotFound):
-            Request._get_method([cat], 'cat_says')
-
-    def test_dict(self):
-        def cat(): pass
-        def dog(): pass
-        dictionary = {'cat_says': cat, 'dog_says': dog}
-        self.assertIs(cat, Request._get_method(dictionary, 'cat_says'))
-        self.assertIs(dog, Request._get_method(dictionary, 'dog_says'))
-
-    def test_dict_non_existant(self):
-        def cat(): pass
-        with self.assertRaises(MethodNotFound):
-            Request._get_method({'cat_says': cat}, 'cat')
-
-    def test_methods_object(self):
-        def cat(): pass
-        def dog(): pass
-        methods = Methods()
-        methods.add(cat)
-        methods.add(dog)
-        self.assertIs(cat, Request._get_method(methods, 'cat'))
-        self.assertIs(dog, Request._get_method(methods, 'dog'))
+    def test_false(self):
+        req = Request({'jsonrpc': '2.0', 'method': 'foo', 'id': 99})
+        self.assertFalse(req.is_notification)
 
 
 class TestCall(TestCase):
@@ -175,86 +154,6 @@ class TestCall(TestCase):
         req = Request({'jsonrpc': '2.0', 'method': 'get_name', 'params':
                        {'name': 'foo'}, 'id': 1})
         self.assertEqual('foo', req.call([get_name])['result'])
-
-
-class TestGetArguments(TestCase):
-
-    def test_none(self):
-        self.assertEqual((None, None), Request._get_arguments(
-            {'jsonrpc': '2.0', 'method': 'foo'}))
-
-    def test_positional(self):
-        self.assertEqual(([2, 3], None), Request._get_arguments(
-            {'jsonrpc': '2.0', 'method': 'foo', 'params': [2, 3]}))
-
-    def test_keyword(self):
-        self.assertEqual((None, {'foo': 'bar'}), Request._get_arguments(
-            {'jsonrpc': '2.0', 'method': 'foo', 'params': {'foo': 'bar'}}))
-
-    def test_invalid_none(self):
-        with self.assertRaises(InvalidParams):
-            Request._get_arguments({'jsonrpc': '2.0', 'method': 'foo', 'params': None})
-
-    def test_invalid_numeric(self):
-        with self.assertRaises(InvalidParams):
-            Request._get_arguments({'jsonrpc': '2.0', 'method': 'foo', 'params': 5})
-
-    def test_invalid_string(self):
-        with self.assertRaises(InvalidParams):
-            Request._get_arguments({'jsonrpc': '2.0', 'method': 'foo', 'params': 'str'})
-
-
-class TestRequestInit(TestCase):
-
-    def tearDown(self):
-        config.convert_camel_case = False
-
-    def test_invalid_request(self):
-        req = Request({'jsonrpc': '2.0'})
-        self.assertIsInstance(req.response, ErrorResponse)
-
-    def test_ok(self):
-        req = Request({'jsonrpc': '2.0', 'method': 'foo'})
-        self.assertEqual('foo', req.method_name)
-
-    def test_positional_args(self):
-        req = Request({'jsonrpc': '2.0', 'method': 'foo', 'params': [2, 3]})
-        self.assertEqual([2, 3], req.args)
-
-    def test_keyword_args(self):
-        req = Request({'jsonrpc': '2.0', 'method': 'foo', 'params': {'foo': 'bar'}})
-        self.assertEqual({'foo': 'bar'}, req.kwargs)
-
-    def test_request_id(self):
-        req = Request({'jsonrpc': '2.0', 'method': 'foo', 'id': 99})
-        self.assertEqual(99, req.request_id)
-
-    def test_request_id_notification(self):
-        req = Request({'jsonrpc': '2.0', 'method': 'foo'})
-        self.assertEqual(None, req.request_id)
-
-    def test_convert_camel_case(self):
-        config.convert_camel_case = True
-        req = Request({'jsonrpc': '2.0', 'method': 'fooMethod', 'params': {
-            'fooParam': 1, 'aDict': {'barParam': 1}}})
-        self.assertEqual('foo_method', req.method_name)
-        self.assertEqual({'foo_param': 1, 'a_dict': {'bar_param': 1}}, req.kwargs)
-
-    def test_positional_args_convert_case_skip(self):
-        config.convert_camel_case = True
-        req = Request({'jsonrpc': '2.0', 'method': 'foo', 'params': ['Camel', 'Case']})
-        self.assertEqual(['Camel', 'Case'], req.args)
-
-
-class TestRequestIsNotification(TestCase):
-
-    def test_true(self):
-        req = Request({'jsonrpc': '2.0', 'method': 'foo'})
-        self.assertTrue(req.is_notification)
-
-    def test_false(self):
-        req = Request({'jsonrpc': '2.0', 'method': 'foo', 'id': 99})
-        self.assertFalse(req.is_notification)
 
 
 class TestRequestProcessNotifications(TestCase):
